+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * (#36) Fix bug with some URLs
 * `title`:
+  - When album lists titles in the quoted form (**Artist "Title"**), split artist from the
+    title correctly.
   - Address a long-standing issue with track names of the form **Title - Some Mix** where we
     would parse **Title** as the artist and **Some Mix** as the title. Such name now gets
     replaced by **Title (Some Mix)** which is then attributed correctly.